requirements.txt
setuptools
requests==2.18.4
selenium==3.5.0
selenium-requests==1.3
git+https://github.com/Kitryn/PTCAccount2.git
unirest
pylint
beautifulsoup4
python-dotenv